Update monkey testing functionality to embed gremlins.js npm package into project
