+++ Lombiq.Tests.UI/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "dependencies": {
+    "gremlins.js": "2.2.0"
+  }
+}
